requirements.txt
# ==============================================================
# DJI AG API - Dependencies
# ==============================================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Browser Automation
playwright>=1.40.0

# HTTP Client
httpx>=0.26.0

# Data Processing
protobuf>=4.25.0
openpyxl>=3.1.0

# Utilities
python-dateutil>=2.8.0
